[rfc] fx quant: add QuantizationFXTraceable marker
Summary:

Adds a `QuantizationFXTraceable` class to mark a subgraph as
symbolically traceable for quantization. Adds top level API
support to `prepare_fx` and `convert_fx` to be able to take
non-traceable models with traceable subgraphs, defined by
this marker.

RFC for now, to check with the team on the API. Once we align
on the API, we can fix the docblocks and add more rigorous testing.

Test Plan:

```
python test/test_quantization.py TestQuantizeFx.test_fx_traceable_marker
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8d2afcc297febbfd8d574f8ee4eef5f7dde91754
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47215