quant fused conv-bn: calculate conv_w rescale without grad
Summary:

While analyzing the long standing issue with gradient explosions
when QAT is trained from scratch, I noticed that we are backpropogating
to `bn.weight` when calculating the scaling factor for a conv-bn's conv
weights.  We should discuss in person, but I'm guessing that this was
not intended. Please see https://arxiv.org/pdf/1806.08342.pdf Figure 9
for details.

In an unrealistic experiment, this change improves the stability of
the conv-bn block, although it does not eliminate it completely.  Setup:
* Create torchvision's quantizeable MobileNetV2 with random weights
* fuse it
* set LR to 1.0 (smaller values also lead to a repro, 1.0 is convenient for a fast repro, although not very realistic).
* measure the # of iterations from start of training to seeing the first NaNs
* the number ^ is ~20 iterations without this PR, and increases to ~100 iterations with this PR
* (more analysis is FB-only for now, TODO write it up)

Note: I've traced the further instability to the interplay between bn's
running variance and the conv weight scaling, we can discuss a separate
recommendation for that outside of this PR.

Note: only the `with torch.no_grad()` line is for commit,
the rest is debugging info which is convenient to have during analysis
and will be removed before landing.

Test Plan:

See above.
TODO: experiments on a more realistic setting

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2bbbb477b5716f4c5be6317a3885a1584f0a5308
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49981