[JIT] Add buffer/parameter metadata test to test_save_load.py
**Summary**
This commit adds a unit test to `test_save_load.py` that checks that
saving and loading a module preserves metadata about which module
attributes are parameters and buffers. The hooks that are currently used
to automatically check serialization of every function and module in the
unit tests check that the archive produced by saving and loading and
saving again are the same and that the type tags for the actual IValues
representing the module match before saving and after loading. However,
these tests do not check that buffer and parameter metadata was not
lost or destroyed during serialization.

**Test Plan**
Ran the new unit test.

ghstack-source-id: 96b11981f9ef7ef1cad38c6c2c1e6034454a5681
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49594